\name{spatstat.data-package}
\alias{spatstat.data-package} 
\alias{spatstat.data} 
\docType{package}
\title{The spatstat.data Package}
\description{
  The \pkg{spatstat.data} package contains the datasets
  for the \pkg{spatstat} family of packages.
}
\details{
  The \pkg{spatstat.data} package contains the datasets
  for the \pkg{spatstat} family of packages.

  These are spatial datasets; they are objects belonging to
  classes of spatial data defined in other sub-packages
  of the \pkg{spatstat} family.
  In order to handle these datasets correctly, we recommend
  loading the \pkg{spatstat} package.
}
\section{Licence}{
  This library and its documentation are usable under the terms of the
  \dQuote{GNU General Public License},
  a copy of which is distributed with \R.
}
\author{
  \spatstatAuthors.
}
\keyword{spatial}
\keyword{package}
