\name{spatstat.data-package}
\alias{spatstat.data-package} 
\alias{spatstat.data} 
\docType{package}
\title{The spatstat.data Package}
\description{
  The \pkg{spatstat.data} package contains the datasets
  for the \pkg{spatstat} package.
}
\details{
  The \pkg{spatstat.data} package contains the datasets
  for the \pkg{spatstat} package.
}
\section{Licence}{
  This library and its documentation are usable under the terms of the
  \dQuote{GNU General Public License},
  a copy of which is distributed with \R.
}
\author{
  \spatstatAuthors.
}
\keyword{spatial}
\keyword{package}
